Add files via upload
The code to create a SQL file including datetime column, and my_variable column for specific day with time interval of 1 minute. This code will create the SQL file from beginning and it will fill the columns automatically. Additionally each of the parameters can be modified based on the user need.